crud op for DeliveryAgent

+++ src/main/java/jsp/courier/repository/DeliveryAgentRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import jsp.courier.entity.DeliveryAgent;
 
-	public interface DeliveryAgentRepository extends JpaRepository<DeliveryAgent, Integer> {
+	public interface DeliveryAgentRepository extends JpaRepository<DeliveryAgent, Long> {
 
 
 }
